Lots of code from ETQW
Lexer/Parser update.
New idLib containers and some useful classes.
...some prec.header mess...

+++ neo/idlib/math/Simd.h
@@ -92,7 +92,7 @@ public:
 		cpuid = CPUID_NONE;
 	}
 	
-	cpuid_t							cpuid;
+	cpuid_t	cpuid;
 	
 	virtual const char* VPCALL		GetName() const = 0;
 	
@@ -102,8 +102,8 @@ public:
 	virtual	void VPCALL MinMax( idVec3& min,		idVec3& max,			const idDrawVert* src,	const int count ) = 0;
 	virtual	void VPCALL MinMax( idVec3& min,		idVec3& max,			const idDrawVert* src,	const triIndex_t* indexes,		const int count ) = 0;
 	
-	virtual void VPCALL Memcpy( void* dst,			const void* src,		const int count ) = 0;
-	virtual void VPCALL Memset( void* dst,			const int val,			const int count ) = 0;
+	virtual void VPCALL Memcpy( void* dst,			const void* src,		const size_t count ) = 0;
+	virtual void VPCALL Memset( void* dst,			const int val,			const size_t count ) = 0;
 	
 	// animation
 	virtual void VPCALL BlendJoints( idJointQuat* joints, const idJointQuat* blendJoints, const float lerp, const int* index, const int numJoints ) = 0;
